refactor(app): split chat window events
